refactor(playlist): split refine sheet from detail view
